code cleaned, example updated and legacy code removed
