+++ src/argilla_llama_index/llama_index_handler.py
@@ -5,6 +5,7 @@ from packaging.version import parse
 from typing import Any, Dict, List, Optional
 import warnings
 
+import argilla as rg
 from argilla._constants import DEFAULT_API_KEY, DEFAULT_API_URL
 from llama_index.core.callbacks.base_handler import BaseCallbackHandler
 from llama_index.core.callbacks.schema import (
@@ -130,6 +131,17 @@ class ArgillaCallbackHandler(BaseCallbackHandler):
                 )
 
                 self.dataset = dataset.push_to_argilla(self.dataset_name)
+                self.is_new_dataset_created = True
+                warnings.warn(
+                    (
+                        f"No dataset with the name {self.dataset_name} was found in workspace "
+                        f"{self.workspace_name}. A new dataset with the name {self.dataset_name} "
+                        "has been created with the question fields `prompt` and `response`"
+                        "and the rating question `response-rating` with values 1-7 and text question"
+                        " named `response-feedback`."
+                    ),
+                )
+
             else:
                 # Update the existing dataset. If the fields and questions do not match, the dataset will be updated using the
                 # -updated flag in the name.
@@ -138,6 +150,7 @@ class ArgillaCallbackHandler(BaseCallbackHandler):
                         name=self.dataset_name,
                         workspace=self.workspace_name,
                     )
+
                     self.is_new_dataset_created = False
 
                     if number_of_retrievals > 0:
@@ -199,7 +212,7 @@ class ArgillaCallbackHandler(BaseCallbackHandler):
         except Exception as e:
             raise FileNotFoundError(
                 f"`FeedbackDataset` retrieval and creation both failed with exception `{e}`."
-                f" If the problem persists please report it to {self.ISSUES_URL} "
+                f" If the problem persists please report it to https://github.com/argilla-io/argilla/issues/ "
                 f"as an `integration` issue."
             ) from e
 
@@ -225,7 +238,362 @@ class ArgillaCallbackHandler(BaseCallbackHandler):
         self.components_to_log = set()
         self.event_ids_traced = set()
 
-        
+    def _add_context_fields(self, number_of_retrievals: int) -> List:
+        """Create the context fields to be added to the dataset."""
+        context_fields = [
+            rg.TextField(
+                name="retrieved_document_" + str(doc + 1),
+                title="Retrieved Document " + str(doc + 1),
+                use_markdown=True,
+                required=False,
+            )
+            for doc in range(number_of_retrievals)
+        ]
+        return context_fields
+
+    def _add_context_questions(self, number_of_retrievals: int) -> List:
+        """Create the context questions to be added to the dataset."""
+        rating_questions = [
+            rg.RatingQuestion(
+                name="rating_retrieved_document_" + str(doc + 1),
+                title="Rate the relevance of the Retrieved Document "
+                + str(doc + 1)
+                + " (if present)",
+                values=list(range(1, 8)),
+                # After https://github.com/argilla-io/argilla/issues/4523 is fixed, we can use the description
+                description=None,  # "Rate the relevance of the retrieved document."
+                required=False,
+            )
+            for doc in range(number_of_retrievals)
+        ]
+        return rating_questions
+
+    def _create_root_and_other_nodes(self, trace_map: Dict[str, List[str]]) -> None:
+        """Create the root node and the other nodes in the tree."""
+        self.root_node = self._get_event_name_by_id(trace_map["root"][0])
+        self.event_ids_traced = set(trace_map.keys()) - {"root"}
+        print(trace_map)
+        self.event_ids_traced.update(*trace_map.values())
+        for id in self.event_ids_traced:
+            self.components_to_log.add(self._get_event_name_by_id(id))
+
+    def _get_event_name_by_id(self, event_id: str) -> str:
+        """Get the name of the event by its id."""
+        return str(self.events_data[event_id][0].event_type).split(".")[1].lower()
+
+        # TODO: If we have a component more than once, properties currently don't account for those after the first one and get overwritten
+
+    def _add_missing_metadata_properties(
+        self,
+        dataset: rg.FeedbackDataset,
+    ) -> None:
+        """Add missing metadata properties to the dataset."""
+        required_metadata_properties = []
+        for property in self.components_to_log:
+            metadata_name = f"{property}_time"
+            if property == self.root_node:
+                metadata_name = "total_time"
+            required_metadata_properties.append(metadata_name)
+
+        existing_metadata_properties = [
+            property.name for property in dataset.metadata_properties
+        ]
+        missing_metadata_properties = [
+            property
+            for property in required_metadata_properties
+            if property not in existing_metadata_properties
+        ]
+
+        for property in missing_metadata_properties:
+            title = " ".join([word.capitalize() for word in property.split("_")])
+            if title == "Llm Time":
+                title = "LLM Time"
+            dataset.add_metadata_property(
+                rg.FloatMetadataProperty(name=property, title=title)
+            )
+            if self.is_new_dataset_created == False:
+                warnings.warn(
+                    (
+                        f"The dataset given was missing some required metadata properties. "
+                        f"Missing properties were {missing_metadata_properties}. "
+                        f"Properties have been added to the dataset with "
+                    ),
+                )
+
+    def _check_components_for_tree(
+        self, tree_structure_dict: Dict[str, List[str]]
+    ) -> Dict[str, List[str]]:
+        """
+        Check whether the components in the tree are in the components to log.
+        Removes components that are not in the components to log so that they are not shown in the tree.
+        """
+        final_components_in_tree = self.components_to_log.copy()
+        final_components_in_tree.add("root")
+        for component in self._ignore_components_in_tree:
+            if component in final_components_in_tree:
+                final_components_in_tree.remove(component)
+        for key in list(tree_structure_dict.keys()):
+            if key.strip("0") not in final_components_in_tree:
+                del tree_structure_dict[key]
+        for key, value in tree_structure_dict.items():
+            if isinstance(value, list):
+                tree_structure_dict[key] = [
+                    element
+                    for element in value
+                    if element.strip("0") in final_components_in_tree
+                ]
+        return tree_structure_dict
+
+    def _get_events_map_with_names(
+        self, events_data: Dict[str, List[CBEvent]], trace_map: Dict[str, List[str]]
+    ) -> Dict[str, List[str]]:
+        """
+        Returns a dictionary where trace_map is mapped with the event names instead of the event ids.
+        Also returns a set of the event ids that were traced.
+        """
+        self.event_map_id_to_name = {}
+        for event_id in self.event_ids_traced:
+            event_name = str(events_data[event_id][0].event_type).split(".")[1].lower()
+            while event_name in self.event_map_id_to_name.values():
+                event_name = event_name + "0"
+            self.event_map_id_to_name[event_id] = event_name
+        events_trace_map = {
+            self.event_map_id_to_name.get(k, k): [
+                self.event_map_id_to_name.get(v, v) for v in values
+            ]
+            for k, values in trace_map.items()
+        }
+
+        print("Events trace map with names:")
+        print(events_trace_map)
+
+        return events_trace_map
+
+    def _extract_and_log_info(
+        self, events_data: Dict[str, List[CBEvent]], trace_map: Dict[str, List[str]]
+    ) -> None:
+        """
+        Main function that extracts the information from the events and logs it to Argilla.
+        We currently log data if the root node is either "agent_step" or "query".
+        Otherwise, we do not log anything.
+        If we want to account for more root nodes, we just need to add them to the if statement.
+        """
+        print(f"Events data:")
+        for key, value in events_data.items():
+            print(f"{key}: {value}")
+        print(f"Trace map {trace_map}")
+        events_trace_map = self._get_events_map_with_names(events_data, trace_map)
+        root_node = trace_map["root"]
+        data_to_log = {}
+
+        if len(root_node) == 1:
+
+            # Create logging data for the root node
+            if self.root_node == "agent_step":
+                # Event start
+                event = events_data[root_node[0]][0]
+                data_to_log["query"] = event.payload.get(EventPayload.MESSAGES)[0]
+                query_start_time = event.time
+                # Event end
+                event = events_data[root_node[0]][1]
+                data_to_log["response"] = event.payload.get(
+                    EventPayload.RESPONSE
+                ).response
+                query_end_time = event.time
+                data_to_log["agent_step_time"] = _get_time_diff(
+                    query_start_time, query_end_time
+                )
+
+            elif self.root_node == "query":
+                # Event start
+                event = events_data[root_node[0]][0]
+                data_to_log["query"] = event.payload.get(EventPayload.QUERY_STR)
+                query_start_time = event.time
+                # Event end
+                event = events_data[root_node[0]][1]
+                data_to_log["response"] = event.payload.get(
+                    EventPayload.RESPONSE
+                ).response
+                query_end_time = event.time
+                data_to_log["query_time"] = _get_time_diff(
+                    query_start_time, query_end_time
+                )
+
+            else:
+                return
+
+            # Create logging data for the rest of the components
+            self.event_ids_traced.remove(root_node[0])
+            number_of_components_used = defaultdict(int)
+            components_to_log_without_root_node = self.components_to_log.copy()
+            components_to_log_without_root_node.remove(self.root_node)
+            retrieval_metadata = {}
+            for id in self.event_ids_traced:
+                event_name = self.event_map_id_to_name[id]
+                if event_name.endswith("0"):
+                    event_name_reduced = event_name.strip("0")
+                    number_of_components_used[event_name_reduced] += 1
+                else:
+                    event_name_reduced = event_name
+
+                for component in components_to_log_without_root_node:
+                    if event_name_reduced == component:
+                        data_to_log[f"{event_name}_time"] = _calc_time(events_data, id)
+
+                if event_name_reduced == "llm":
+                    data_to_log[f"{event_name}_system_prompt"] = (
+                        events_data[id][0].payload.get(EventPayload.MESSAGES)[0].content
+                    )
+                    data_to_log[f"{event_name}_model_name"] = events_data[id][
+                        0
+                    ].payload.get(EventPayload.SERIALIZED)["model"]
+
+                retrieved_document_counter = 1
+                if event_name_reduced == "retrieve":
+                    for retrieval_node in events_data[id][1].payload.get(
+                        EventPayload.NODES
+                    ):
+                        retrieve_dict = retrieval_node.to_dict()
+                        retrieval_metadata[
+                            f"{event_name}_document_{retrieved_document_counter}_score"
+                        ] = retrieval_node.score
+                        retrieval_metadata[
+                            f"{event_name}_document_{retrieved_document_counter}_filename"
+                        ] = retrieve_dict["node"]["metadata"]["file_name"]
+                        retrieval_metadata[
+                            f"{event_name}_document_{retrieved_document_counter}_text"
+                        ] = retrieve_dict["node"]["text"]
+                        retrieval_metadata[
+                            f"{event_name}_document_{retrieved_document_counter}_start_character"
+                        ] = retrieve_dict["node"]["start_char_idx"]
+                        retrieval_metadata[
+                            f"{event_name}_document_{retrieved_document_counter}_end_character"
+                        ] = retrieve_dict["node"]["end_char_idx"]
+                        retrieved_document_counter += 1
+                        if retrieved_document_counter > self.number_of_retrievals:
+                            break
+
+            metadata_to_log = {}
+
+            for keys in data_to_log.keys():
+                if keys == "query_time" or keys == "agent_step_time":
+                    metadata_to_log["total_time"] = data_to_log[keys]
+                elif keys.endswith("_time"):
+                    metadata_to_log[keys] = data_to_log[keys]
+                elif keys != "query" and keys != "response":
+                    metadata_to_log[keys] = data_to_log[keys]
+
+            if len(number_of_components_used) > 0:
+                for key, value in number_of_components_used.items():
+                    metadata_to_log[f"number_of_{key}_used"] = value + 1
+
+            metadata_to_log.update(retrieval_metadata)
+
+            tree_structure = self._create_tree_structure(events_trace_map, data_to_log)
+            tree = self._create_svg(tree_structure)
+
+            fields = {
+                "prompt": data_to_log["query"],
+                "response": data_to_log["response"],
+                "time-details": tree,
+            }
+
+            if self.number_of_retrievals > 0:
+                for key, value in list(retrieval_metadata.items()):
+                    if key.endswith("_text"):
+                        fields[f"retrieved_document_{key[-6]}"] = (
+                            f"DOCUMENT SCORE: {retrieval_metadata[key[:-5]+'_score']}\n\n"
+                            + value
+                        )
+                        del metadata_to_log[key]
+
+            self.dataset.add_records(
+                records=[
+                    {"fields": fields, "metadata": metadata_to_log},
+                ]
+            )
+
+    def _create_tree_structure(
+        self, events_trace_map: Dict[str, List[str]], data_to_log: Dict[str, Any]
+    ) -> List:
+        """Create the tree data to be converted to an SVG."""
+        events_trace_map = self._check_components_for_tree(events_trace_map)
+        data = []
+        data.append(
+            (
+                0,
+                0,
+                self.root_node.strip("0").upper(),
+                data_to_log[f"{self.root_node}_time"],
+            )
+        )
+        current_row = 1
+        for root_child in events_trace_map[self.root_node]:
+            data.append(
+                (
+                    current_row,
+                    1,
+                    root_child.strip("0").upper(),
+                    data_to_log[f"{root_child}_time"],
+                )
+            )
+            current_row += 1
+            for child in events_trace_map[root_child]:
+                data.append(
+                    (
+                        current_row,
+                        2,
+                        child.strip("0").upper(),
+                        data_to_log[f"{child}_time"],
+                    )
+                )
+                current_row += 1
+        return data
+
+    def _create_svg(self, data: List) -> str:
+        """
+        Create an SVG file from the data.
+
+        Args:
+            data (List): The data to create the SVG file from.
+
+        Returns:
+            str: The SVG file.
+        """
+
+        # changing only the box height changes all other values as well
+        # others can be adjusted individually if needed
+        box_height = 47
+        box_width = box_height * 8.65
+        row_constant = box_height + 7
+        indent_constant = 40
+        font_size_node_name = box_height * 0.4188
+        font_size_time = font_size_node_name - 4
+        text_centering = box_height * 0.6341
+        node_name_indent = box_height * 0.35
+        time_indent = box_height * 7.15
+
+        body = ""
+        for each in data:
+            row, indent, node_name, node_time = each
+            body_raw = f"""
+<g transform="translate({indent*indent_constant}, {row*row_constant})">
+<rect x=".5" y=".5" width="{box_width}" height="{box_height}" rx="8.49" ry="8.49" style="fill: #24272e; stroke: #afdfe5; stroke-miterlimit: 10;"/>
+<text transform="translate({node_name_indent} {text_centering})" style="fill: #fff; font-size: {font_size_node_name}px;"><tspan x="0" y="0">{node_name}</tspan></text>
+<text transform="translate({time_indent} {text_centering})" style="fill: #b7d989; font-size: {font_size_time}px; font-style: italic;"><tspan x="0" y="0">{node_time}</tspan></text>
+</g>
+            """
+            body += body_raw
+            base = (
+                base
+            ) = f"""
+<?xml version="1.0" encoding="UTF-8"?>
+<svg id="Layer_1" data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 750 {len(data)*row_constant}">
+{body}
+</svg>
+            """
+            base = base.strip()
+        return base
 
     # The four methods required by the abstrac class BaseCallbackHandler.
     # These methods are the one being executed on the different events, by the llama-index
@@ -263,18 +631,9 @@ class ArgillaCallbackHandler(BaseCallbackHandler):
 
         self._trace_map = trace_map or defaultdict(list)
         self._end_time = datetime.now()
-        print("Events data on end_trace()")
-        for key, value in self.events_data.items():
-            print(key, value)
-        print("Trace map on end_trace()")
-        print(self._trace_map)
-        print()
-
-        # The trace_map is a dictionary with the event_id as the key and the list of components as the value. However,
-        # it only register those events which payload is of type EventPayload.QUERY_STR. We must manually introduce
-        # the events that are not of this type, but are relevant to the prediction process.
-        # for node_id in self._trace_map:
-        #     for event_id, event_id_object in self.events_data.items():
+        self._create_root_and_other_nodes(trace_map)
+        self._add_missing_metadata_properties(self.dataset)
+        self._extract_and_log_info(self.events_data, trace_map)
 
     def on_event_start(
         self,
@@ -318,40 +677,40 @@ class ArgillaCallbackHandler(BaseCallbackHandler):
 
         event = CBEvent(event_type, payload=payload, id_=event_id)
         self.events_data[event_id].append(event)
-        self._trace_map = defaultdict(list)
 
-    # Auxiliary methods
 
-    def _get_time_diff(event_1_time_str: str, event_2_time_str: str) -> float:
-        """
-        Get the time difference between two events Follows the American format (month, day, year).
+# Auxiliary methods
+def _get_time_diff(event_1_time_str: str, event_2_time_str: str) -> float:
+    """
+    Get the time difference between two events Follows the American format (month, day, year).
 
-        Args:
-            event_1_time_str (str): The first event time.
-            event_2_time_str (str): The second event time.
+    Args:
+        event_1_time_str (str): The first event time.
+        event_2_time_str (str): The second event time.
 
-        Returns:
-            float: The time difference between the two events.
-        """
-        time_format = "%m/%d/%Y, %H:%M:%S.%f"
+    Returns:
+        float: The time difference between the two events.
+    """
+    time_format = "%m/%d/%Y, %H:%M:%S.%f"
 
-        event_1_time = datetime.strptime(event_1_time_str, time_format)
-        event_2_time = datetime.strptime(event_2_time_str, time_format)
+    event_1_time = datetime.strptime(event_1_time_str, time_format)
+    event_2_time = datetime.strptime(event_2_time_str, time_format)
 
-        return round((event_2_time - event_1_time).total_seconds(), 4)
+    return round((event_2_time - event_1_time).total_seconds(), 4)
 
-    def _calc_time(events_data: Dict[str, List[CBEvent]], id: str) -> float:
-        """
-        Calculate the time difference between the start and end of an event using the events_data.
 
-        Args:
-            events_data (Dict[str, List[CBEvent]]): The events data, stored in a dictionary.
-            id (str): The event id to calculate the time difference between start and finish timestamps.
+def _calc_time(events_data: Dict[str, List[CBEvent]], id: str) -> float:
+    """
+    Calculate the time difference between the start and end of an event using the events_data.
 
-        Returns:
-            float: The time difference between the start and end of the event.
-        """
+    Args:
+        events_data (Dict[str, List[CBEvent]]): The events data, stored in a dictionary.
+        id (str): The event id to calculate the time difference between start and finish timestamps.
 
-        start_time = events_data[id][0].time
-        end_time = events_data[id][1].time
-        return _get_time_diff(start_time, end_time)
+    Returns:
+        float: The time difference between the start and end of the event.
+    """
+
+    start_time = events_data[id][0].time
+    end_time = events_data[id][1].time
+    return _get_time_diff(start_time, end_time)
